Consolidate tone
