--- inodaqv2/command_tone.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "Arduino.h"
-
-namespace Command
-{
-void command_tone(const ::String &command);
-}
