+++ inodaqv2/commands.hpp
@@ -7,3 +7,4 @@ void command_dig(const ::String &command);
 void command_dread();
 void command_ping();
 void command_pwm(const ::String &command);
+void command_tone(const ::String &command);
